Allow the hamburger menu button to be left aligned
Previously we only allowed the hamburger button to be in the minor
modes section of the mode line.  That meant the button was fairly
right, but annoyingly not completely aligned to that side.

Aligning the button, to either side, is good for consistency with most
hamburger menu buttons out there.

Thanks to Jamie Nicol for this feature request.